Move HasExpireDate implementation to trait
